[fix]
Fixed the problems in the profile and profile editor pages
