+++ lib/presenter/my_profile/my_profile_page.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/material.dart';
+
+import 'my_profile_editor/profile_editor_page.dart';
+import 'widgets/body_my_profile.dart';
+
+class MyProfilePage extends StatelessWidget {
+  const MyProfilePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        elevation: 0,
+        centerTitle: true,
+        toolbarHeight: 80,
+        backgroundColor: Color.fromARGB(255, 99, 66, 191),
+        title: const Padding(
+          padding: EdgeInsets.only(left: 10),
+          child: Text(
+            'Minha Conta',
+            style: TextStyle(fontSize: 30),
+          ),
+        ),
+        leading: IconButton(
+          icon: Icon(
+            Icons.arrow_back_rounded,
+            size: 35,
+          ),
+          onPressed: () {
+            Navigator.pop(context);
+          },
+        ),
+        actions: [
+          Padding(
+            padding: const EdgeInsets.only(right: 10),
+            child: IconButton(
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (context) {
+                      return ProfileEditorPage();
+                    },
+                  ),
+                );
+              },
+              icon: const Icon(
+                Icons.edit_note,
+                size: 35,
+              ),
+            ),
+          ),
+        ],
+      ),
+      body: BodyMyProfile(),
+    );
+  }
+}
