+++ lib/presenter/my_profile/widgets/close_button_profile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../login/login.dart';
+import '../../login/login_page.dart';
 
 class CloseButtonProfile extends StatelessWidget {
   const CloseButtonProfile({
@@ -61,7 +61,7 @@ class CloseButtonProfile extends StatelessWidget {
                   Navigator.of(context).push(
                     MaterialPageRoute(
                       builder: (context) {
-                        return Login();
+                        return LoginPage();
                       },
                     ),
                   );
