+++ lib/presenter/splash_screen/splash_screen_page.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 
-import '../login/login.dart';
+import '../login/login_page.dart';
 
-class SplashScreen extends StatefulWidget {
-  const SplashScreen({Key? key}) : super(key: key);
+class SplashScreenPage extends StatefulWidget {
+  const SplashScreenPage({Key? key}) : super(key: key);
 
   @override
-  State<SplashScreen> createState() => _SplashScreenState();
+  State<SplashScreenPage> createState() => _SplashScreenPageState();
 }
 
-class _SplashScreenState extends State<SplashScreen> {
+class _SplashScreenPageState extends State<SplashScreenPage> {
   @override
   void initState() {
     Future.delayed(
@@ -18,7 +18,7 @@ class _SplashScreenState extends State<SplashScreen> {
       Navigator.of(context).push(
         MaterialPageRoute(
           builder: (context) {
-            return Login();
+            return LoginPage();
           },
         ),
       );
